source/requirements.txt
asgiref==3.6.0
autopep8==2.0.2
black==23.3.0
boto3==1.28.11
botocore==1.31.11
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
Django==4.2.1
django-cors-headers==3.14.0
django-extensions==3.2.1
django-filter==23.2
django-seed==0.3.1
django-storages==1.13.2
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.2
djangorestframework-word-filter==0.2.0
Faker==19.3.0
gunicorn==21.2.0
idna==3.4
jmespath==1.0.1
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.1
Pillow==9.5.0
platformdirs==3.5.1
psycopg==3.1.9
psycopg2-binary==2.9.6
pycodestyle==2.10.0
pydotplus==2.0.2
PyJWT==1.7.1
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
requests==2.30.0
s3transfer==0.6.1
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
toposort==1.10
typing_extensions==4.5.0
urllib3==1.26.16
